+++ CHANGELOG.md
@@ -5,6 +5,13 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/), and this project adheres
 to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## [1.7.1] - 2025-06-13
+
+### Changed
+
+- Avoid recording duplicate actions
+- Limit the amount of recorded reactions to a certain amount for resilience reasons
+
 ## [1.7.0] - 2025-06-13
 
 ### Changed